Extract code into separate modules
Why:
- Functions are easier to find
- Future childminder side implementation will be easier

I only moved code around, there's not any other change.
So as soon as eslint is happy, it should be safe.
